moving get_test_options to a diff file
